Added logic to pass outputs to frontend

+++ src-tauri/src/events/output.rs
@@ -1,5 +1,4 @@
 use serde::{Deserialize, Serialize};
-use simconnect::DWORD;
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct Output {
@@ -7,6 +6,6 @@ pub struct Output {
     pub metric: String,
     pub update_every: f32,
     pub cb_text: String,
-    pub id: DWORD,
-    pub output_type: DWORD,
+    pub id: u32,
+    pub output_type: i8
 }
